Support output devices
Removed Write from CaptureDevice, moved to new OutputDevice
Refactored some things shared between Capture and Output out of capture
