+++ src/output/parameters.rs
@@ -1,39 +1,13 @@
-use bitflags::bitflags;
 use std::{fmt, mem};
 
 use crate::fraction::Fraction;
 use crate::parameters::Capabilities;
 use crate::v4l_sys::*;
 
-bitflags! {
-    pub struct Modes: u32 {
-        const HIGH_QUALITY      = 0x1000;
-    }
-}
-
-impl From<u32> for Modes {
-    fn from(caps: u32) -> Self {
-        Modes::from_bits_truncate(caps)
-    }
-}
-
-impl Into<u32> for Modes {
-    fn into(self) -> u32 {
-        self.bits()
-    }
-}
-
-impl fmt::Display for Modes {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt::Debug::fmt(self, f)
-    }
-}
-
 #[derive(Debug, Copy, Clone)]
-/// Streaming parameters (single-planar)
+/// Output parameters (single-planar)
 pub struct Parameters {
     pub capabilities: Capabilities,
-    pub modes: Modes,
     pub interval: Fraction,
 }
 
@@ -41,7 +15,6 @@ impl Parameters {
     pub fn new(frac: Fraction) -> Self {
         Parameters {
             capabilities: Capabilities::from(0),
-            modes: Modes::from(0),
             interval: frac,
         }
     }
@@ -49,7 +22,6 @@ impl Parameters {
     pub fn with_fps(fps: u32) -> Self {
         Parameters {
             capabilities: Capabilities::from(0),
-            modes: Modes::from(0),
             interval: Fraction::new(1, fps),
         }
     }
@@ -58,31 +30,28 @@ impl Parameters {
 impl fmt::Display for Parameters {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         writeln!(f, "capabilities : {}", self.capabilities)?;
-        writeln!(f, "modes        : {}", self.modes)?;
         writeln!(f, "interval     : {} [s]", self.interval)?;
         Ok(())
     }
 }
 
-impl From<v4l2_captureparm> for Parameters {
-    fn from(params: v4l2_captureparm) -> Self {
+impl From<v4l2_outputparm> for Parameters {
+    fn from(params: v4l2_outputparm) -> Self {
         Parameters {
             capabilities: Capabilities::from(params.capability),
-            modes: Modes::from(params.capturemode),
             interval: Fraction::from(params.timeperframe),
         }
     }
 }
 
-impl Into<v4l2_captureparm> for Parameters {
-    fn into(self: Parameters) -> v4l2_captureparm {
-        let mut params: v4l2_captureparm;
+impl Into<v4l2_outputparm> for Parameters {
+    fn into(self: Parameters) -> v4l2_outputparm {
+        let mut params: v4l2_outputparm;
         unsafe {
             params = mem::zeroed();
         }
 
         params.capability = self.capabilities.into();
-        params.capturemode = self.modes.into();
         params.timeperframe = self.interval.into();
         params
     }
